fix(PropertyValue): change propertyVAL type from int to double and update related methods

+++ src/main/java/com/log/dao/PropertyValuesDAO.java
@@ -12,7 +12,7 @@ public class PropertyValuesDAO {
         String sql = "INSERT into Property_Values (Prop_VAL, Property_ID) VALUES (?,?)";
 
         try(PreparedStatement stmt = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)){
-            stmt.setInt(1, propertyValue.getPropertyVAL());
+            stmt.setDouble(1, propertyValue.getPropertyVAL());
             stmt.setInt(2, propertyValue.getPropertyID());
             stmt.executeUpdate();
             ResultSet rs = stmt.getGeneratedKeys();
