+++ src/main/java/com/log/model/PropertyValue.java
@@ -2,16 +2,16 @@ package com.log.model;
 
 public class PropertyValue {
     private int propertyValID;
-    private int propertyVAL;
+    private double propertyVAL;
     private int propertyID;
 
-    public PropertyValue(int propertyValID, int propertyVAL, int propertyID) {
+    public PropertyValue(int propertyValID, double propertyVAL, int propertyID) {
         this.propertyValID = propertyValID;
         this.propertyVAL = propertyVAL;
         this.propertyID = propertyID;
     }
 
-    public PropertyValue(int propertyVAL, int propertyID) {
+    public PropertyValue(double propertyVAL, int propertyID) {
         this.propertyVAL = propertyVAL;
         this.propertyID = propertyID;
     }
@@ -24,11 +24,11 @@ public class PropertyValue {
         this.propertyValID = propertyValID;
     }
 
-    public int getPropertyVAL() {
+    public double getPropertyVAL() {
         return propertyVAL;
     }
 
-    public void setPropertyVAL(int propertyVAL) {
+    public void setPropertyVAL(double propertyVAL) {
         this.propertyVAL = propertyVAL;
     }
